+++ R/plot.R
@@ -39,7 +39,7 @@ mc_plot_loggers <- function(data, directory, localities=NULL, sensors=NULL, crop
     oldpar <- par(no.readonly = TRUE)
     on.exit(suppressWarnings(par(oldpar)))
     .common_stop_if_not_raw_format(data)
-    data <- mc_filter(data, localities, sensors)
+    data <- mc_filter(data, localities=localities, sensors=sensors)
     .prep_check_datetime_step_unprocessed(data)
     loggers <- .common_get_loggers(data)
     dir.create(directory, showWarnings = F)